Middleware admin

+++ routes/web.php
@@ -31,14 +31,29 @@ Auth::routes();
 
 Route::get('/home', 'HomeController@index')->name('home');
 
+
+//Route::get('admin/vehicules', 'VehiculesController@index')->name('adminvehicules');
+//
+//Route::get('admin/vehicules/create', 'VehiculesController@create');
+//Route::post('admin/vehicules/create', 'VehiculesController@store');
+//
+//Route::get('admin/vehicules/edit/{id}', 'VehiculesController@edit');
+//Route::post('admin/vehicules/edit/{id}', 'VehiculesController@update');
+//
+//Route::delete('admin/vehicules/delete/{id}', 'VehiculesController@destroy');
+
+
+Route::group([
+    'middleware' => ['auth', 'admin']
+], function () {
 // Véhicules
-Route::get('vehicules', 'VehiculesController@index')->name('vehicules');
+Route::get('admin/vehicules', 'VehiculesController@index')->name('adminvehicules');
 
-Route::get('vehicules/create', 'VehiculesController@create');
-Route::post('vehicules/create', 'VehiculesController@store');
+Route::get('admin/vehicules/create', 'VehiculesController@create');
+Route::post('admin/vehicules/create', 'VehiculesController@store');
 
-Route::get('vehicules/edit/{id}', 'VehiculesController@edit');
-Route::post('vehicules/edit/{id}', 'VehiculesController@update');
+Route::get('admin/vehicules/edit/{id}', 'VehiculesController@edit');
+Route::post('admin/vehicules/edit/{id}', 'VehiculesController@update');
 
-Route::delete('vehicules/delete/{id}', 'VehiculesController@destroy');
-
+Route::delete('admin/vehicules/delete/{id}', 'VehiculesController@destroy');
+});
